fix(embedding): a MESMA base fp16 derrubou o treino pela segunda vez, em outro arquivo
O T1f morreu aos 2 min com `ValueError: Attempting to unscale FP16 gradients`.

`thenlper/gte-base` guarda os pesos em fp16 e o `transformers` novo carrega no
dtype do checkpoint. O AMP exige pesos-mestres em fp32 — é ele que faz a passagem
em fp16, e o `GradScaler` que desescala os gradientes de volta; um modelo já em
fp16 não tem para onde desescalar.

## O que faz este commit valer mais que a linha que ele muda

**Isto já tinha acontecido.** Em 2026-09-03 o braço `gte` do T1c morreu com o mesmo
erro, com o mesmo modelo. Foi diagnosticado, escrito numa nota de 10 linhas em
`kaggle/t1c_phirank.py`, e consertado — **em `rerank.py`**.

`embedding.py` nunca recebeu a correção, porque nenhuma base fp16 tinha passado por
ele: MiniLM e SciBERT são fp32, então o defeito era invisível aqui. A lição estava
paga, documentada, e aplicada num arquivo só.

O teste novo cobre os DOIS módulos de treino de uma vez, por AST, e exige também
que a razão esteja escrita junto — `dtype=torch.float32` parece redundante (fp32 é
"o padrão") e é exatamente o tipo de coisa que o próximo removedor de redundância
apaga sem saber o que está apagando.

## O que o run entregou antes de morrer

O GTE-base zero-shot no pool interno de 1000 candidatos: recall@1 0,475 · recall@10
0,812 · MRR 0,591 · **nDCG@10 0,639**. É a linha de base do braço, medida de graça.

E a precaução da célula funcionou: `kernels output` devolveu o log da API com **0
bytes** de novo, e o `treino_t1f.log` gravado em /kaggle/working veio com 63 KB —
foi dele que saiu o traceback.

Custo: ~2 min de cota.

789 testes.

+++ src/phifm/training/embedding.py
@@ -351,8 +351,28 @@ class TreinadorEmb:
         # materializar a matriz N×N é justamente o tensor de 226 MB que matou o
         # treino de lote 128 duas vezes nesta máquina.
         self.atencao = "sdpa" if self.dev.type == "cuda" else "eager"
+        # ⚠️ `dtype=torch.float32` EXPLÍCITO, e o custo de não ter foi medido duas
+        # vezes, no mesmo repositório, com o mesmo modelo.
+        #
+        # `thenlper/gte-base` guarda os pesos em fp16, e o `transformers` novo
+        # carrega no dtype do checkpoint por padrão. O AMP exige pesos-mestres em
+        # fp32 — é ele que faz a passagem em fp16, e o `GradScaler` que desescala
+        # os gradientes de volta. Um modelo já em fp16 não tem para onde
+        # desescalar:
+        #
+        #     ValueError: Attempting to unscale FP16 gradients.
+        #
+        # Isto foi diagnosticado e consertado em `rerank.py` em 2026-09-03, quando
+        # custou um braço inteiro do T1c. A correção ficou LÁ. Este módulo nunca a
+        # recebeu porque nenhuma base fp16 tinha passado por ele — o MiniLM e o
+        # SciBERT são fp32, então o defeito era invisível aqui.
+        #
+        # Em 2026-09-11 o T1f morreu aos 2 min com o mesmo erro: a lição estava
+        # paga, escrita e aplicada num arquivo só. Não é propriedade do gte —
+        # qualquer checkpoint fp16 quebraria.
         self.mod = AutoModel.from_pretrained(
-            cfg.base, attn_implementation=self.atencao).to(self.dev)
+            cfg.base, attn_implementation=self.atencao,
+            dtype=torch.float32).to(self.dev)
         # AMP só faz sentido onde há suporte de verdade. O DirectML não expõe
         # `GradScaler` utilizável, e ligar autocast lá deu queda silenciosa no
         # backward — então a decisão é pelo dispositivo, não por bandeira do
